uint

+++ spectests/structs_encoding.go
@@ -121,7 +121,7 @@ func (e *ETHMergeTransactions) HashTreeRootWith(hh *ssz.Hasher) (err error) {
 		for i := uint64(0); i < num; i++ {
 			txSubIndx := hh.Index()
 			hh.PutBytes(e.OpaqueList[i])
-			hh.MerkleizeWithMixin(txSubIndx, len(e.OpaqueList[i]), 1048576)
+			hh.MerkleizeWithMixin(txSubIndx, uint64(len(e.OpaqueList[i])), 1048576)
 		}
 		hh.MerkleizeWithMixin(subIndx, num, 16384)
 	}
